// packages/shared/src/meta/generated/InputField.generated.ts
export const INPUT_FIELD_PROPS_SCHEMA = z.object({
    invalid: z.boolean().optional(),
    labelSlot: z.string().optional(),
    label: z.string().optional(),
    description: z.string().optional(),
    infoIcon: z.boolean().optional(),
    infoIconSlot: z.string().optional(),
    infoTooltipContent: z.string().optional(),
    infoIconAriaLabel: z.string().optional(),
    fullWidth: z.boolean().optional(),
    error: z.string().optional(),
    feedback: z.string().optional(),
    dynamicTextSlot: z.string().optional(),
    dynamicText: z.string().optional(),
    helperButton: z.string().optional(),
    validationMode: z.enum(["onBlur", "onChange"]).optional(),
    validate: z.any().optional(),
    start: z.string().optional(),
    end: z.string().optional(),
    disabled: z.boolean().optional(),
    name: z.string().optional(),
    type: z.string().optional(),
    value: z.string().optional(),
    defaultValue: z.string().optional(),
    autoFocus: z.boolean().optional(),
    id: z.string().optional(),
    "aria-label": z.string().optional(),
    onFocus: z.any().optional(),
    onBlur: z.any().optional(),
    onChange: z.any().optional(),
    onKeyDown: z.any().optional(),
    size: z.union([z.literal("xs"), z.literal("s"), z.literal("m"), z.literal("l"), z.literal(6), z.literal(8), z.literal(10), z.literal(12)]).optional(),
    appearance: z.enum(["auto", "primary", "secondary", "neutral", "sparkle", "positive", "negative", "warning", "informative"]).optional(),
    shape: z.enum(["default", "pill"]).optional(),
    attention: z.enum(["high", "medium"]).optional(),
    start2: z.string().optional(),
    end2: z.string().optional(),
    placeholder: z.string().optional(),
    readOnly: z.boolean().optional(),
    required: z.boolean().optional(),
    "data-testid": z.string().optional(),
    autoComplete: z.string().optional(),
    maxLength: z.number().optional()
  }).strict();